+++ clients/admin-ui/src/features/datastore-connections/system_portal_config/TestConnectionMessage.tsx
@@ -3,21 +3,24 @@ import React from "react";
 
 type TestConnectionMessageProps = {
   status: "success" | "error";
+  failure_reason?: string;
 };
 
-const TestConnectionMessage = ({ status }: TestConnectionMessageProps) => (
-  <>
-    {status === "success" && (
-      <Text data-testid="toast-success-msg">
-        Connection test was successful
-      </Text>
-    )}
-    {status === "error" && (
-      <Text data-testid="toast-error-msg">
-        Test failed: please check your connection info
-      </Text>
-    )}
-  </>
-);
+const TestConnectionMessage = ({
+  status,
+  failure_reason,
+}: TestConnectionMessageProps) => {
+  if (status === "error") {
+    let description = "Connection test failed.";
+    if (failure_reason) {
+      description += ` ${failure_reason}`;
+    }
+    return <Text data-testid="toast-error-msg">{description}</Text>;
+  }
+
+  return (
+    <Text data-testid="toast-success-msg">Connection test was successful</Text>
+  );
+};
 
 export default TestConnectionMessage;