+++ src/fides/api/schemas/connection_configuration/enums/google_cloud_sql_ip_type.py
@@ -1,0 +1,9 @@
+from enum import Enum
+
+
+class GoogleCloudSQLIPType(str, Enum):
+    """Enum for Google Cloud SQL IP types"""
+
+    public = "public"
+    private = "private"
+    private_service_connect = "psc"