+++ src/fides/api/tasks/storage.py
@@ -134,7 +134,9 @@ def upload_to_s3(  # pylint: disable=R0913
         s3_client = get_s3_client(
             auth_method,
             storage_secrets,
-            assume_role_arn=CONFIG.credentials["storage"].get("aws_s3_assume_role_arn"),
+            assume_role_arn=CONFIG.credentials.get(  # pylint: disable=no-member
+                "storage", {}
+            ).get("aws_s3_assume_role_arn"),
         )
 
         # handles file chunking